ci: shard live functional tests
