+++ e2e/live/test_responses.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from e2e.live.helpers import create_client
+
+
+def test_responses() -> None:
+    with create_client() as client:
+        response = client.responses.create(
+            input="Reply with only the word pong.",
+            max_output_tokens=16,
+            preset="pro-search",
+        )
+
+    assert response.id
+    assert response.object == "response"
+    assert response.output
+    assert response.status == "completed"
